Add simulation subsystems, game features, and multiplayer protocol
- anno-sim: AI controller, combat system, trade routes, population model,
  A* pathfinding, carrier dispatch, warehouse inventory, marketplace
  coverage, ocean pathfinding, data bridge from COD/SZS
- anno-formats: COL palette parser, SZS scenario parser, enhanced COD
  parser with sub-object properties
- anno-render: SpriteManager for all 33 BSH files across 3 zoom levels
- anno-audio: StreamManager auto-detect finished tracks
- anno-game: game binary with building placement, sound integration,
  sprite Y-offset fix; island-viewer and sim-test binaries
- anno-net: new crate — Maxnet.dll decompiled (77 functions via Ghidra),
  DirectPlay protocol reverse-engineered (12 message types), TCP
  transport replacement with host/client architecture

+++ crates/anno-sim/src/production.rs
@@ -34,12 +34,12 @@ pub fn tick_building(building: &mut BuildingInstance, def: &BuildingDef, dt_ms: 
     // Advance production timer
     building.production_timer_ms += dt_ms;
 
-    if building.production_timer_ms < def.cycle_time_ms as u32 {
+    if building.production_timer_ms < def.cycle_time_ms {
         return 0;
     }
 
     // Production cycle complete
-    building.production_timer_ms -= def.cycle_time_ms as u32;
+    building.production_timer_ms -= def.cycle_time_ms;
 
     // Consume inputs
     if def.input_1_rate > 0 {
